Convert sass to scss
Scss syntax can be more terse because it allows collapsing multiple
attributes into a single line.  This is particularly important for
animations, so we surrender.  It is sad that this is the case, sass
could otherwise include optional braces, but...

This change also does a bit of cleanup, removing several dead selectors.