Fix artifacts when scaling the sinusoid
When scaling the sinusoid, sometimes, a white line would appear around
its border.  This seems to be caused because when layering a
`background-image` and a `background-color`, the background-image would
not completely cover the colored area.  We now produce manually the
stacking, by drawing the sinusoid image in a nested `::after`
pseudoelement.  This pseudoelement is 1px bigger to make sure it always
overlaps the white area underneath.