Unfortunate combination of refactorings
It all started with me trying to extract a new `slideshow-view` from the
`detail-view`, that could be reused in in the `think-view` and others.
But Reagent has several ways one can communicate with an inner
component, so I went into a refactoring frenzy which ended like this.

The main problem I was facing is how to communicate was is the current
url and previous url to the slideshow.  Innitially I was tempted to have
the slideshow keep track of this, but this was essentially a bad idea.
The only source of truth about the current navigation position is the
browser URL, and it should be forwarded to the application state in a
single, simple, route handler in `routes/`.

This commit removes the hack of having ad-hoc manipulation of the
state (e.g path `:do :detail`) in route handlers, we just pass the
current `:view` to the components that need it.  We reassemble the shape
of the state as needed using `track`, see for example the instantiation
of the `slideshow/view`.

Another big refactoring is that the state has been modularized -- every
module providing a stateful Reagent component should now provide also a
default `state`.  These components are expected to be required with `:as
my-component` and provide a `my-component/view` and
`my-component/state`.  Parent components can then define their own state
in term of it's childrens, and component modules encapsulate all the
information needed to use them.

There are also some other unneceesary style changes here.  One notable
is the avoidance of `(letfn [ ... ] (r/with-let [ ... ]))` and just use
the `with-let` to define the functions.