Only one sound can play at time in /think
This is achieved by decoupling the player view from the actual player
object with an underlying `<audio>` tag.  Only one player object is
used, that is shared accross all player views.

This has the drawback that play position and buffering is not persistent
anymore accross plays.  It can be restored later though.

The big advantage is that by decoupling player views from player logic,
it is easy now to add a new player view that is seamlessly synchronized
in the detail and thumnail view.