Replace showdown by markdown-it
The later seems to be more complete and have better plugins.

In this commit we also change the way we handle packaging JavaScript
libraries.  Instead of adding them individualy to `foreign-libs`, we
bundle them together with `browserify` and add this bundle to the cljs
build.  This means we can use `js/require` in ClojureScript, and extern
definition becomes greatly simplified.

We also tried directly referencing the libraries from npm with
`:commonjs` in `:foreign-libs`, but it did not work:

    http://dev.clojure.org/jira/browse/CLJS-1646